+++ tests/test_dials_data.py
@@ -44,7 +44,9 @@ def test_datafetcher_constructs_py_path(fetcher, root):
         ds = df("dataset")
     assert pathlib.Path(ds).resolve() == pathlib.Path("/tmp/root/dataset").resolve()
     assert isinstance(ds, py.path.local)
-    fetcher.assert_called_once_with("dataset", pre_scan=True, read_only=False)
+    fetcher.assert_called_once_with(
+        "dataset", pre_scan=True, read_only=False, verify=True
+    )
 
     ds = df("dataset", pathlib=False)
     assert pathlib.Path(ds).resolve() == pathlib.Path("/tmp/root/dataset").resolve()
@@ -64,7 +66,9 @@ def test_datafetcher_constructs_path(fetcher, root):
     assert ds == test_path / "dataset"
 
     assert isinstance(ds, pathlib.Path)
-    fetcher.assert_called_once_with("dataset", pre_scan=True, read_only=False)
+    fetcher.assert_called_once_with(
+        "dataset", pre_scan=True, read_only=False, verify=True
+    )
 
     with pytest.warns(DeprecationWarning):
         ds = df("dataset")
@@ -72,4 +76,6 @@ def test_datafetcher_constructs_path(fetcher, root):
     assert not isinstance(
         ds, pathlib.Path
     )  # default is currently to return py.path.local()
-    fetcher.assert_called_once_with("dataset", pre_scan=True, read_only=False)
+    fetcher.assert_called_once_with(
+        "dataset", pre_scan=True, read_only=False, verify=True
+    )